DCAC-78: Rationalise integration tests following merge.
